# Avoid non-API calls when truncating overallocated vectors