--- a/cpp11test/R/cpp11.R
+++ b/cpp11test/R/cpp11.R
@@ -231,3 +231,7 @@ test_destruction_inner <- function() {
 test_destruction_outer <- function() {
   invisible(.Call(`_cpp11test_test_destruction_outer`))
 }
+
+cpp11_push_and_truncate_ <- function(size_sexp) {
+  .Call(`_cpp11test_cpp11_push_and_truncate_`, size_sexp)
+}