--- a/cpp11test/src/truncate.cpp
+++ b/cpp11test/src/truncate.cpp
@@ -3,15 +3,11 @@
 [[cpp11::register]] SEXP cpp11_push_and_truncate_(SEXP size_sexp) {
   R_xlen_t size = INTEGER(size_sexp)[0];
 
+  // Allocate `size` worth of doubles (filled with garbage data)
   cpp11::writable::doubles out(size);
 
-  // Fill it
-  for (R_xlen_t i = 0; i < size; ++i) {
-    out.push_back(0);
-  }
-
-  // Push 1 more past the existing capacity,
-  // doubling the capacity
+  // Push 1 more past the existing length/capacity,
+  // doubling the capacity for cpp11 vectors
   out.push_back(0);
 
   // Truncate back to `size + 1` size and return result.