--- a/vignettes/motivations.Rmd
+++ b/vignettes/motivations.Rmd
@@ -487,7 +487,8 @@ b_grow <- bench::press(.grid = grid,
   {
     fun = match.fun(sprintf("%sgrow_", ifelse(pkg == "cpp11", "", paste0(pkg, "_"))))
     bench::mark(
-      fun(len)
+      fun(len),
+      min_iterations = 100
     )
   }
 )[c("len", "pkg", "min", "mem_alloc", "n_itr", "n_gc")]