@pachadotdev 2024-08-19
excellent!

here is some meta text to make this issue easier to find with Google, Bing, etc

If CRAN checks reveal messages such as

```
gcc.exe (GCC) 13.2.0
File 'mypkg/libs/x64/mypkg.dll':
  Found non-API calls to R: 'SETLENGTH', 'SET_GROWABLE_BIT',
    'SET_TRUELENGTH'
```

Or:

```
Debian clang version 18.1.8 (9)
File 'mypkg/libs/x64/mypkg.so':
  Found non-API calls to R: 'SETLENGTH', 'SET_GROWABLE_BIT',
    'SET_TRUELENGTH'
```

this is the fix :)