@pachadotdev 2024-08-21
Hi @DavisVaughan 
Do you have plans to send this to CRAN soon? I cannot update a pkg I have hosted there because of these warnings. If the update should take long, I can vendor cpp11 headers from this PR and solve it.