@DavisVaughan 2024-08-21
> I cannot update a pkg I have hosted there because of these warnings

You can. CRAN knows about them and will ignore them if you tell them they come from cpp11. They have done that for many of us.

I do not have an exact release date yet. Still working on some fixes and minor features. But soon ish, probably a 1 month timeline depending on breakages